Move all capnp related code to Transition-Backend
Only Transition as a need to use capnp file, so we move the all the base file from Chaire-Lib-Backend to Transition-Backend

We move the files from the original path src/services/json2capnp/ into existing directory where they were used.
