+++ packages/chaire-lib-backend/package.json
@@ -42,7 +42,6 @@
         "JSONStream": "^1.3.5",
         "bcryptjs": "^3.0.3",
         "bytes": "^3.1.2",
-        "capnp-ts": "^0.4.0",
         "chaire-lib-common": "^0.2.2",
         "dotenv": "^17.2.3",
         "express": "^5.2.1",
