+++ packages/transition-backend/package.json
@@ -40,6 +40,7 @@
         "express": "^5.2.1",
         "express-session": "^1.18.2",
         "express-socket.io-session": "^1.3.5",
+        "fetch-retry": "^6.0.0",
         "geobuf": "^3.0.2",
         "geojson": "^0.5.0",
         "glob": "^13.0.0",
